ci: Add caching
Adds caching where the push and periodic workflows both restore and
save the cache. The PR workflow only restores the cache. The cache size
is expected to be ~9 GiB, which is under the 10 GiB limit. If the cache
size becomes too large, it can be quickly disabled, but left in the code
to be re-enabled once increased cache sizes are available.
(See https://github.com/github/roadmap/issues/1029)

Signed-off-by: Colin Pinnell McAllister <colinmca242@gmail.com>