ci: Fix getting build targets
The paths output by changed-files are paths relative to the
meta-lts-collab repository root, which is not equivalent to
GITHUB_WORKSPACE in the bitbake workflow. I'm not sure how to get
changed-files to use absolute paths, so the easy fix is to modify the
targets step to run the get-bitbake-targets command from the base
directory of meta-lts-collab.

Signed-off-by: Colin Pinnell McAllister <colinmca242@gmail.com>