ci: Always save cache for push and periodic
Updates push-periodic workflow to always save the cache. This allows
push events to both use and save a new instance of the cache. GitHub
should automatically replace the old cache with the new cache due to
cache size. This will allow push events to update the cache over the
week, reducing build times for push and pull request workflows.

Signed-off-by: Colin Pinnell McAllister <colinmca242@gmail.com>